Enhanced exception handling in OnApplicationStarted method
In the `CronDaemonService.cs` file, modified the `OnApplicationStarted` method. Moved the `var jobsToRun = _cronDaemon.CronJobs.Values` statement to a `try` block and added a `try-catch` structure to catch and handle possible exceptions. In the `try` block, check if `jobsToRun` has any jobs to run, and if so, log and run those jobs. In the `catch` block, log any exceptions that occur in the `OnApplicationStarted` method. These changes improve the robustness of the code and ensure that exceptions are properly handled when running jobs at startup.